// src/app/feature/components/inv-eq/interface/inv-eq.interface.ts
// @Autor:Gianni Martinez
export interface IInvEq {
  readonly id: string;
  readonly registration: string;
  readonly typeEq: string;
  readonly trademark: string;
  readonly model: string;
  readonly serial: string;
  readonly actualSit: string;
  readonly state: string;
  readonly entity: string;
}